HH-218702 move OpenTelemetryJdbcExtension to nab-jdbc-common module
