+++ nab-data-source/pom.xml
@@ -20,6 +20,11 @@
         </dependency>
 
         <dependency>
+            <groupId>ru.hh.nab</groupId>
+            <artifactId>nab-jdbc-common</artifactId>
+        </dependency>
+
+        <dependency>
             <groupId>jakarta.annotation</groupId>
             <artifactId>jakarta.annotation-api</artifactId>
         </dependency>
