+++ nab-data-source/src/main/java/ru/hh/nab/datasource/NabDataSourceProdConfig.java
@@ -5,9 +5,9 @@ import jakarta.inject.Named;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import static ru.hh.nab.common.qualifier.NamedQualifier.SERVICE_NAME;
-import ru.hh.nab.datasource.ext.OpenTelemetryJdbcExtension;
 import ru.hh.nab.datasource.healthcheck.HealthCheckHikariDataSourceFactory;
 import ru.hh.nab.datasource.monitoring.NabMetricsTrackerFactoryProvider;
+import ru.hh.nab.jdbc.common.ext.OpenTelemetryJdbcExtension;
 import ru.hh.nab.metrics.StatsDSender;
 
 @Configuration
